requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-annotate requirements.in
#
asgiref==3.7.2
brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dj-database-url==2.1.0
django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
idna==3.6
packaging==23.2
psycopg2==2.9.9
pycountry==23.12.11
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==2.1.0
whitenoise[brotli]==6.6.0
